Changes to be committed: 	modified:   floflis 	modified:   init.sh 	modified:   install.sh
Fixed to fit the FHS (https://en.wikipedia.org/wiki/Filesystem_Hierarchy_Standard)
Removed redundant sudo.

+++ init.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 ok="Ok, going to the next question."
+invalid="Please enter a valid input"
 
 read -r -d '' dna_ascii << "EOF"
 
@@ -23,6 +24,7 @@ echo "Installing without git makes it lightweight (ideal for routers and other l
 echo "Update module and .git folder haves approx. 349.2 KB of disk usage."
 echo "Do you want to use git to update Floflis? (Alternative is a manual update)"
 echo "[Y/n]"
+while true; do
 read gitormanual
 
 case gitormanual in
@@ -31,11 +33,16 @@ case gitormanual in
       sudo rm -r .git
       echo "Removing update module..."
       sudo rm modules/update
-      echo "$ok" ;;
+      echo "${ok}"
+      break ;;
    
+   [yY])
+       echo "${ok}"
+       break ;;
    *)
-      echo "$ok" ;;
+       echo "${invalid}" ;;
 esac
+done
 echo "${dna_ascii}"
 
 echo "Do you know how to turn off/restart/suspend Linux?"
@@ -43,15 +50,20 @@ echo "You can have the convenience to do it with 'floflis shu/sus/res' commands.
 echo "But, if you don't want the 1,5 KB of the Floflis' energy module and knows how to do it from Linux:"
 echo ""
 echo "Use energy module? [Y/n]"
+while true; do
 read keepenergymod
 case $keepenergymod in
    [nN])
        echo "Removing energy module..."
        rm modules/energy
+       echo "${ok}"
+       break ;;
+   [yY])
        echo "${ok}" ;;
    *)
-       echo "${ok}" ;;
+       echo "${invalid}";;
 esac
+done
 echo "${dna_ascii}"
 
 echo "Floflis DNA contains README.md, LICENSE, CONTRIBUTING, DISCLAIMER and CODE_OF_CONDUCT."
@@ -61,7 +73,7 @@ echo "Documentation and information-reader module haves approx. 6.6 KB of disk u
 echo "Keep documentation? [Y/n]"
 read keepdocumentation
 
-# if [ "$keepdocumentation" = "y" ]; then
+# if [ "$keepdocumentation" = "y" ]; then # No reason to keep this, doesnt do anything useful.
 #   echo "${ok}"
 # fi
 echo "${dna_ascii}"
@@ -73,21 +85,26 @@ echo "Before removing documentation..."
 echo "Do you agree with the license and the disclaimer? Please scroll up to read."
 echo "PLEASE READ AND WRITE CAREFULLY!"
 echo "Do you agree to the license? [Y/n]"
+while true; do
 read licenseagreement
 case $licenseagreement in
    [nN])
       echo "You need to agree to the license in order to use Floflis!"
-      exit ;;
-   *)
+      exit;;
+   [yY])
       echo "Removing README.md" && rm README.md
       echo "Removing LICENSE file" && rm LICENSE
       echo "Removing CONTRIBUTING file" && rm CONTRIBUTING
       echo "Removing DISCLAIMER file" && rm DISCLAIMER
       echo "Removing CODE_OF_CONDUCT.md" && rm CODE_OF_CONDUCT.md
-      echo "Removing information-reader module" && rm modules/information-reader ;;
+      echo "Removing information-reader module" && rm modules/information-reader
+      break ;;
+   *)
+      echo "${invalid}" ;;
 esac
+done
    
 echo "${dna_ascii}"
 
 echo "Done! Now, you can move this folder to the device you want to install Floflis, and run install.sh using 'sudo chmod +x install.sh && sudo ./install.sh' command (without quotes)."
-sudo rm init.sh
+# rm init.sh # Usually, wizards like this one should not remove themselves.
